requirements.txt
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
pyobjc-framework-Quartz>=10.0
rumps>=0.4.0
slack_sdk>=3.21.0
